+++ src/main/java/blackjack/view/ResultView.java
@@ -3,6 +3,7 @@ package blackjack.view;
 import blackjack.domain.card.Hand;
 import blackjack.domain.card.Suit;
 import blackjack.domain.participant.Participant;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -24,7 +25,7 @@ public final class ResultView {
     private static final String CARD_SCORE_FORMAT = "%s카드: %s - 결과: %d";
 
     private static final String TITLE_PROFITS = "## 최종 수익";
-    private static final String PROFITS_FORMAT = "%s: %d";
+    private static final String PROFITS_FORMAT = "%s: %.0f";
 
     public void dealInitialCards(final List<String> playerNames, final Entry<String, Hand> dealerCards,
                                  final Map<String, Hand> playersCards) {
@@ -68,7 +69,7 @@ public final class ResultView {
                 .collect(Collectors.joining(COMMA));
     }
 
-    public void showProfit(final Map<Participant, Integer> profits) {
+    public void showProfit(final Map<Participant, BigDecimal> profits) {
         showLine(LINE + TITLE_PROFITS);
         profits.entrySet().stream()
                 .map(entry -> String.format(PROFITS_FORMAT, entry.getKey().getNickname(), entry.getValue()))
